@@ -1,6 +1,6 @@
 {
   "name": "tauri",
-  "version": "0.2.1",
+  "version": "0.3.1",
   "description": "Multi-binding collection of libraries and templates for building Tauri apps",
   "bin": {
     "tauri": "./bin/tauri.js"
@@ -55,10 +55,9 @@
     "imagemin-pngquant": "8.0.0",
     "imagemin-zopfli": "6.0.0",
     "is-png": "2.0.0",
-    "isbinaryfile": "4.0.2",
+    "isbinaryfile": "4.0.3",
     "jsdom": "15.2.1",
-    "lodash.debounce": "4.0.8",
-    "lodash.template": "4.5.0",
+    "lodash": "4.17.15",
     "minimist": "1.2.0",
     "ms": "2.1.2",
     "png2icons": "2.0.1",
@@ -86,11 +85,11 @@
     "babel-jest": "^24.9.0",
     "cors": "^2.8.5",
     "dotenv": "8.2.0",
-    "eslint": "6.7.2",
-    "eslint-config-standard-with-typescript": "^11.0.1",
+    "eslint": "6.8.0",
+    "eslint-config-standard-with-typescript": "11.0.1",
     "eslint-plugin-import": "2.19.1",
     "eslint-plugin-lodash-template": "0.15.0",
-    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-node": "11.0.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-security": "1.4.0",
     "eslint-plugin-standard": "4.0.1",
@@ -100,12 +99,12 @@
     "jest-mock-process": "1.2.0",
     "jest-raw-loader": "^1.0.1",
     "lint-staged": "9.5.0",
-    "lockfile-lint": "3.0.3",
+    "lockfile-lint": "3.0.5",
     "promise": "8.0.3",
-    "raw-loader": "^4.0.0",
+    "raw-loader": "4.0.0",
     "ts-loader": "6.2.1",
-    "typescript": "3.7.3",
-    "webpack": "4.41.4",
+    "typescript": "3.7.4",
+    "webpack": "4.41.5",
     "webpack-cli": "3.3.10",
     "webpack-node-externals": "1.7.2"
   },
