@@ -17,15 +17,15 @@
     "tauri:prod:build": "yarn tauri:prod build"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^6.0.0",
-    "rollup": "^1.12.0",
-    "rollup-plugin-commonjs": "^10.0.0",
-    "rollup-plugin-livereload": "^1.0.0",
-    "rollup-plugin-svelte": "^5.0.3",
-    "rollup-plugin-terser": "^5.1.2",
-    "sirv-cli": "^0.4.4"
+    "@rollup/plugin-node-resolve": "7.0.0",
+    "rollup": "1.29.0",
+    "rollup-plugin-commonjs": "10.1.0",
+    "rollup-plugin-livereload": "1.0.4",
+    "rollup-plugin-svelte": "5.1.1",
+    "rollup-plugin-terser": "5.1.3",
+    "sirv-cli": "0.4.5"
   },
   "dependencies": {
-    "svelte": "^3.0.0"
+    "svelte": "3.16.7"
   }
 }
