@@ -22,20 +22,20 @@
   },
   "dependencies": {
     "@quasar/extras": "1.3.3",
-    "quasar": "1.5.9"
+    "quasar": "1.6.1"
   },
   "devDependencies": {
     "@quasar/app": "1.4.3",
-    "@tauri-apps/tauri-webpack": "0.1.4",
+    "@tauri-apps/tauri-webpack": "0.2.0",
     "@vue/eslint-config-standard": "5.0.1",
     "babel-eslint": "10.0.3",
-    "eslint": "6.7.2",
+    "eslint": "6.8.0",
     "eslint-loader": "3.0.3",
     "eslint-plugin-import": "2.19.1",
-    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-node": "11.0.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-standard": "4.0.1",
-    "eslint-plugin-vue": "6.0.1",
+    "eslint-plugin-vue": "6.1.2",
     "tauri": "file:../../../cli/tauri.js"
   },
   "engines": {
