@@ -6,9 +6,10 @@ use handlebars::Handlebars;
 use lazy_static::lazy_static;
 
 use std::collections::BTreeMap;
-use std::fs::write;
+use std::fs::{File, write};
 use std::path::PathBuf;
 use std::process::{Command, Stdio};
+use std::io::Write;
 
 // Create handlebars template for shell scripts
 lazy_static! {
@@ -50,11 +51,18 @@ pub fn bundle_project(settings: &Settings) -> crate::Result<Vec<PathBuf>> {
   // write the scripts
   write(&bundle_sh, bundle_temp).or_else(|e| Err(e.to_string()))?;
 
+  // copy seticon binary
+  let seticon = include_bytes!("templates/seticon");
+  let seticon_out = &output_path.join("seticon");
+  let mut seticon_buffer = File::create(seticon_out).or_else(|e| Err(e.to_string()))?;
+  seticon_buffer.write_all(seticon).or_else(|e| Err(e.to_string()))?;
+
   // chmod script for execution
 
   Command::new("chmod")
     .arg("777")
     .arg(&bundle_sh)
+    .arg(&seticon_out)
     .current_dir(output_path)
     .stdout(Stdio::piped())
     .stderr(Stdio::piped())
